requirements.txt
# Data collection & manipulation
pandas>=2.0.0
numpy>=1.24.0
yfinance>=0.2.31
fredapi>=0.5.1

# Technical analysis
ta>=0.11.0

# Machine learning
scikit-learn>=1.3.0
xgboost>=2.0.0
lightgbm>=4.0.0
optuna>=3.4.0

# Deep learning (MLP)
torch>=2.1.0

# Clustering & regime detection
hmmlearn>=0.3.0

# Explainability
shap>=0.43.0

# Statistical tests
scipy>=1.11.0
statsmodels>=0.14.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.13.0
plotly>=5.18.0

# Web framework
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6

# Serialization & config
joblib>=1.3.0
pyyaml>=6.0.1

# Jupyter
jupyter>=1.0.0
ipykernel>=6.25.0
